Retrieved and filtered posts that are similar to a user's past interactions; put backend functions in util files
